+++ config/splat.hd.tt_000.yaml
@@ -39,5 +39,6 @@ segments:
       - [0x10AC, .rodata, 10E8] # func_80172120
       - [0x10C4, .rodata, 10E8] # func_80174038
       - [0x10E8, c, 10E8]
+      - [0x45C4, c, 45E4]
       - [0x48B8, .bss, 10E8]
   - [19852]
