+++ config/splat.pspeu.tt_000.yaml
@@ -42,6 +42,7 @@ segments:
     subalign: 8
     subsegments:
       - [0x80, c, 10E8]
+      - [0x48C8, c, 45E4]
       - [0x4C80, data]
       - [0x5E00, data] # rodata, not data
       - {type: bss, vram: 0x092ed480}
