+++ config/splat.us.tt_000.yaml
@@ -39,5 +39,6 @@ segments:
       - [0x10AC, .rodata, 10E8] # func_80172120
       - [0x10C4, .rodata, 10E8] # func_80174038
       - [0x10E8, c]
+      - [0x45E4, c]
       - [0x48D8, .bss, 10E8]
   - [0xA000]
